implement authentication and role management with JWT support

+++ Backend-spring/src/main/java/com/comp/web/repository/RoleRepository.java
@@ -1,14 +1,13 @@
 package com.comp.web.repository;
 
-import java.util.Optional;
-
+import com.comp.web.model.entity.ERole;
+import com.comp.web.model.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.comp.web.models.ERole;
-import com.comp.web.models.Role;
+import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
-  Optional<Role> findByName(ERole name);
+public interface RoleRepository extends JpaRepository<Role, Integer> {
+    Optional<Role> findByName(ERole name);
 }
